Move splinter errors into separate modules
