+++ daemon/src/splinter/event/error.rs
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2019-2021 Cargill Incorporated
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * -----------------------------------------------------------------------------
+ */
+
+#[derive(Debug)]
+pub struct ScabbardEventConnectionError(pub String);
+
+impl std::error::Error for ScabbardEventConnectionError {}
+
+impl std::fmt::Display for ScabbardEventConnectionError {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "unable to create connection to {}", self.0)
+    }
+}
